feat: increment active sequence flows for each child instance
We need to increment the active sequence flows a correct amount of time. For each active "token" in a flowscope we need to increment once.

E.g.:
Start event -> Service task -> End event. We want to start at the service task. We need to increment the active sequence flows of the PROCESS element instance 1 time.

Start event -> Subprocess -> Start event -> Service task -> End event -> End event. We want to start at the service task. We need to increment the active sequence flows of the PROCESS element instance 1 time (for the subprocess). We need to incremenet the element instance of the subprocess 1 time (for the service task).

If there are multiple starti nstruction (e.g. in the event of a parallel gateway). We need to increment the flowscope containing these elements once per start instruction.